// types/user.d.ts
export interface User {
  id: string;
  userName: string;
  email: string;
  roles?: string[];
  scopes?: string[];
}